No task associated, fixed homepage test which failed now because of chanes made to homepage
