--- Tests/Frontend-Tests/test_homepage_frontend.py
@@ -1,90 +1,0 @@
-# Generated by Selenium IDE
-import pytest
-import time
-import json
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.support import expected_conditions
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
-
-class TestHomepage():
-  def setup_method(self, method):
-    self.driver = webdriver.Chrome()
-    self.vars = {}
-  
-  def teardown_method(self, method):
-    self.driver.quit()
-  
-  def test_homepage(self):
-    # Test name: homepage
-    # Step # | name | target | value | comment
-    # 1 | open | /login |  | 
-    self.driver.get("http://sayless.azurewebsites.net/login")
-    # 2 | setWindowSize | 1552x840 |  | 
-    self.driver.set_window_size(1552, 840)
-    # 3 | click | id=emailInput |  | 
-    self.driver.find_element(By.ID, "emailInput").click()
-    # 4 | type | id=emailInput | gptburc@gmail.com | 
-    self.driver.find_element(By.ID, "emailInput").send_keys("gptburc@gmail.com")
-    # 5 | type | id=passwordInput | password1234 | 
-    self.driver.find_element(By.ID, "passwordInput").send_keys("password1234")
-    # 6 | click | css=.submit |  | 
-    self.driver.find_element(By.CSS_SELECTOR, ".submit").click()
-    # 7 | waitForElementPresent | id=openchat | 30000 | 
-    WebDriverWait(self.driver, 30000).until(expected_conditions.presence_of_element_located((By.ID, "openchat")))
-    # 8 | verifyTitle | SayLess - Home |  | 
-    assert self.driver.title == "SayLess - Home"
-    # 9 | verifyText | css=body > h1 | SayLess | 
-    assert self.driver.find_element(By.CSS_SELECTOR, "body > h1").text == "SayLess"
-    # 10 | verifyText | id=openchat | Open Chats | 
-    assert self.driver.find_element(By.ID, "openchat").text == "Open Chats"
-    # 11 | verifyText | id=suggestedchat | Suggested Users | 
-    assert self.driver.find_element(By.ID, "suggestedchat").text == "Suggested Users"
-    # 12 | verifyText | id=startachat | Start a Chat | 
-    assert self.driver.find_element(By.ID, "startachat").text == "Start a Chat"
-    # 13 | verifyText | css=h2:nth-child(2) | rileybur | 
-    assert self.driver.find_element(By.CSS_SELECTOR, "h2:nth-child(2)").text == "rileybur"
-    # 14 | verifyText | css=label | Bio: | 
-    assert self.driver.find_element(By.CSS_SELECTOR, "label").text == "Bio:"
-    # 15 | verifyText | id=bio_text | this is my bio boy | 
-    assert self.driver.find_element(By.ID, "bio_text").text == "this is my bio boy"
-    # 16 | mouseDownAt | css=.ochat | 211.01249885559082,125.07499694824219 | 
-    element = self.driver.find_element(By.CSS_SELECTOR, ".ochat")
-    actions = ActionChains(self.driver)
-    actions.move_to_element(element).click_and_hold().perform()
-    # 17 | mouseMoveAt | css=.ochat | 211.01249885559082,125.07499694824219 | 
-    element = self.driver.find_element(By.CSS_SELECTOR, ".ochat")
-    actions = ActionChains(self.driver)
-    actions.move_to_element(element).perform()
-    # 18 | mouseUpAt | css=.ochat | 211.01249885559082,125.07499694824219 | 
-    element = self.driver.find_element(By.CSS_SELECTOR, ".ochat")
-    actions = ActionChains(self.driver)
-    actions.move_to_element(element).release().perform()
-    # 19 | click | css=.fa-cog |  | 
-    self.driver.find_element(By.CSS_SELECTOR, ".fa-cog").click()
-    # 20 | click | css=.fas |  | 
-    self.driver.find_element(By.CSS_SELECTOR, ".fas").click()
-    # 21 | waitForElementPresent | id=d7f054f94cf64bdc4f685942c4120a1766faeda3bcaa800399e4850c08ce77c4 | 30000 | 
-    WebDriverWait(self.driver, 30000).until(expected_conditions.presence_of_element_located((By.ID, "d7f054f94cf64bdc4f685942c4120a1766faeda3bcaa800399e4850c08ce77c4")))
-    # 22 | verifyElementPresent | id=d7f054f94cf64bdc4f685942c4120a1766faeda3bcaa800399e4850c08ce77c4 |  | 
-    elements = self.driver.find_elements(By.ID, "d7f054f94cf64bdc4f685942c4120a1766faeda3bcaa800399e4850c08ce77c4")
-    # 23 | verifyText | id=d7f054f94cf64bdc4f685942c4120a1766faeda3bcaa800399e4850c08ce77c4 | DannyA | 
-    assert self.driver.find_element(By.ID, "d7f054f94cf64bdc4f685942c4120a1766faeda3bcaa800399e4850c08ce77c4").text == "DannyA"
-    # 24 | click | id=d7f054f94cf64bdc4f685942c4120a1766faeda3bcaa800399e4850c08ce77c4 |  | 
-    self.driver.find_element(By.ID, "d7f054f94cf64bdc4f685942c4120a1766faeda3bcaa800399e4850c08ce77c4").click()
-    # 25 | selectFrame | index=0 |  | 
-    self.driver.switch_to.frame(0)
-    # 26 | verifyText | id=header | DannyA | 
-    assert self.driver.find_element(By.ID, "header").text == "DannyA"
-    # 27 | selectFrame | relative=parent |  | 
-    self.driver.switch_to.default_content()
-    # 28 | click | css=.fa-sign-out-alt |  | 
-    self.driver.find_element(By.CSS_SELECTOR, ".fa-sign-out-alt").click()
-    # 29 | verifyTitle | SayLess Login |  | 
-    assert self.driver.title == "SayLess Login"
-    # 30 | close |  |  | 
-    self.driver.close()
-
